Make id BigSerial type so it autogenerates

+++ schema.sql
@@ -1,7 +1,7 @@
 /* Database schema to keep the structure of entire database. */
 
 CREATE TABLE animals (
-    id                 INT,
+    id                 BIGSERIAL,
     name               VARCHAR,
     date_of_birth      DATE,
     escape_attempts    INT,
